# jobtap: prototype job-manager plugin support