--- a/doc/conf.py
+++ b/doc/conf.py
@@ -344,8 +344,14 @@ man_pages = [
     ('man3/idset_encode','idset_encode', 'Convert idset to string and string to idset', [author], 3),
     ('man3/idset_encode','idset_decode', 'Convert idset to string and string to idset', [author], 3),
     ('man3/idset_encode','idset_ndecode', 'Convert idset to string and string to idset', [author], 3),
+    ('man3/flux_jobtap_get_flux','flux_jobtap_get_flux', 'Flux jobtap plugin interfaces', [author], 3),
+    ('man3/flux_jobtap_get_flux','flux_jobtap_service_register', 'Flux jobtap plugin interfaces', [author], 3),
+    ('man3/flux_jobtap_get_flux','flux_jobtap_reprioritize_all', 'Flux jobtap plugin interfaces', [author], 3),
+    ('man3/flux_jobtap_get_flux','flux_jobtap_reprioritize_job', 'Flux jobtap plugin interfaces', [author], 3),
+    ('man3/flux_jobtap_get_flux','flux_jobtap_priority_unavail', 'Flux jobtap plugin interfaces', [author], 3),
     ('man5/flux-config-bootstrap', 'flux-config-bootstrap', 'configure Flux instance bootstrap', [author], 5),
     ('man7/flux-broker-attributes', 'flux-broker-attributes', 'overview Flux broker attributes', [author], 7),
+    ('man7/flux-jobtap-plugins', 'flux-jobtap-plugins', 'overview Flux jobtap plugin API', [author], 7),
 ]
 
 # -- Options for Intersphinx -------------------------------------------------