--- a/doc/test/spell.en.pws
+++ b/doc/test/spell.en.pws
@@ -527,3 +527,8 @@ dirname
 jps
 xargs
 sep
+RPCs
+jobtap
+unavail
+reprioritize
+reprioritization