--- a/src/broker/module.c
+++ b/src/broker/module.c
@@ -617,7 +617,7 @@ module_t *module_add (modhash_t *mh, const char *path)
     int rc;
 
     dlerror ();
-    if (!(dso = dlopen (path, RTLD_NOW | RTLD_LOCAL | FLUX_DEEPBIND))) {
+    if (!(dso = dlopen (path, RTLD_NOW | RTLD_GLOBAL | FLUX_DEEPBIND))) {
         log_msg ("%s", dlerror ());
         errno = ENOENT;
         return NULL;