--- a/src/cmd/Makefile.am
+++ b/src/cmd/Makefile.am
@@ -69,7 +69,8 @@ dist_fluxcmd_SCRIPTS = \
 	flux-mini.py \
 	flux-jobs.py \
 	flux-resource.py \
-	flux-admin.py
+	flux-admin.py \
+	flux-jobtap.py
 
 fluxcmd_PROGRAMS = \
 	flux-terminus \