--- a/src/include/flux/jobtap.h
+++ b/src/include/flux/jobtap.h
@@ -1,0 +1,14 @@
+/************************************************************\
+ * Copyright 2021 Lawrence Livermore National Security, LLC
+ * (c.f. AUTHORS, NOTICE.LLNS, COPYING)
+ *
+ * This file is part of the Flux resource manager framework.
+ * For details, see https://github.com/flux-framework.
+ *
+ * SPDX-License-Identifier: LGPL-3.0
+\************************************************************/
+
+/* Allow in-tree programs to #include <flux/jobtap.h> like out-of-tree would.
+ */
+
+#include "src/modules/job-manager/jobtap.h"