… doc/man3/flux_jobtap_get_flux.rst
… doc/man3/index.rst
… doc/man7/flux-jobtap-plugins.rst
… doc/man7/index.rst
… src/common/libflux/test/plugin.c
… src/common/libflux/test/plugin_bar.c
… src/modules/job-manager/event.c
… src/modules/job-manager/job.c
… src/modules/job-manager/job.h
… src/modules/job-manager/jobtap-internal.h
… src/modules/job-manager/jobtap.c
… src/modules/job-manager/jobtap.h
… src/modules/job-manager/plugins/default.c
… src/modules/job-manager/plugins/hold.c
… src/modules/job-manager/prioritize.c
… src/modules/job-manager/prioritize.h
… src/modules/job-manager/restart.c
… src/modules/job-manager/submit.c
… src/modules/job-manager/urgency.c
… src/modules/sched-simple/sched.c
… t/Makefile.am
… t/job-manager/plugins/args.c
… t/job-manager/plugins/priority-wait.c
… t/job-manager/plugins/random.c
… t/job-manager/plugins/test.c
… t/t2216-job-manager-plugins.t