This is an early prototype of a job manager plugin implementation which could support priority plugins.

There is still a lot of work to do here, including some commit refactoring and implementation of plugin configuration, but I'm posting the work in progress to get feedback on the *interface* itself, in case this is headed in the wrong direction.

#### Abstract

A plugin system for job manager is introduced which uses the standard Flux plugin `flux_plugin_t`. Plugins can register for callbacks to get notified about new jobs (on submission or job manager restart) or any other state transitions, along with a special `priority.set` callback which indicates the job manager is offering to let the plugin calculate a job priority.

Job information is passed to the plugins via the `flux_plugin_arg_t` arguments, reducing the actual exported plugin interface to a minimum.

A plugin can return a special value from the PRIORITY state callback which indicates a priority cannot be calculated immediately. Jobs will stay in the PRIORITY state until the plugin explicitly sets a priority.

An interface to request reprioritization of all jobs is introduced. This causes the `priority.set` callback to be called for all jobs in the PRIORITY and SCHED state.

Plugins have access to the job manager Flux handle and a method to create an RPC endpoint under `job-manager.`, which should allow advanced priority plugins to accomplish asynchronous work to keep internal state up-to-date.

#### Description

This PR introduces a new job manager plugin interface which allows plugins to tap into all job state transitions, as well as optionally provide a priority when the job manager needs to calculate one. Job manager plugins are called "jobtap" plugins to avoid long names in interfaces (like `flux_job_manager_plugin_*`).

The prototype is built using the Flux `flux_plugin_t` standard plugin interface, so a conforming plugin need only define the single entrypoint symbol:
```c 
    int flux_plugin_init (flux_plugin_t *p);
```
from which a plugin can register callbacks based on a topic string glob (this should be familiar to users of job shell plugins), e.g.
```c
     flux_plugin_add_handler (p, "*", cb, NULL);
```
would register callback `cb()` for all possible plugin callback topics.

The possible set of callback topics currently include:

 - `job.new`: the job manager is seeing a job for the first time. This correspoonds to jobs either submitted or being reloaded from from the KVS. In the `job.new` callback, the job state must be explicitly checked (though it is definitely not state NEW, since the job manager does not handle jobs in this state).
 - `job.<state>`: on transition to job state `state`. Note, this will not include all job states. Due to architecture of the job manager the first state plugins will see after `job.new` is `job.priority`.
 - `priority.set`: when the job manager wants to recalculate priority of a job.

Instead of using a scheme where an opaque `job_manager_job_t` type is exposed with accessors for data about the job for which callbacks apply, this prototype instead packs all possible job information into the `flux_plugin_arg_t` passed to the plugin, which allows the plugin to unpack only the necessary data, while also allowing for future flexibility. That is the args are packed as:
```c
         flux_plugin_arg_pack (args,
                              FLUX_PLUGIN_ARG_IN,
                              "{s:O s:I s:i s:i s:i s:I s:f}",
                              "jobspec", job->jobspec_redacted,
                              "id", job->id,
                              "userid", job->userid,
                              "urgency", job->urgency,
                              "state", job->state,
                              "priority", job->priority,
                              "t_submit", job->t_submit);
```
The `jobtap_call()` interface (internal to the job manager) also allows arbitrary other arguments to be passed to the plugin (currently the event that caused the state transition is passed, though I'm not sure there's a use case for it).

One benefit of this approach is that a plugin can unpack just the resources section of the jobspec, e.g.
```c
    flux_plugin_arg_unpack (args, FLUX_PLUGIN_ARG_IN,
                            "{s:{s:o}",
                            "jobspec", "resources", &resources);
```
Plugins may also _modify_ the jobspec as well, which could be useful in some advanced use cases I've seen mentioned recently.

The two most important callback topics for a job priority plugin will be the `job.priority` and `priority.set` callbacks. A priority plugin must provide a priority immediately on both of these callbacks. The priority is returned via the `FLUX_PLUGIN_ARG_OUT` args, e.g.:
```C
        flux_plugin_arg_pack (args, FLUX_PLUGIN_ARG_OUT,
                          "{s:i}",
                          "priority", pri);
```
If in the `job.priority` callback (i.e. entry to the PRIORITY state), the plugin cannot provide a priority immediately, then the special value `FLUX_JOBTAP_PRIORITY_UNAVAIL` should be returned. This will cause the job to stay in the PRIORITY state until the plugin asynchronously provides a priority via the function `flux_jobtap_reprioritize_job()`. A convenience function is provided for this case:
```c
        int flux_jobtap_priority_unavail (flux_plugin_t *p,
                                          flux_plugin_arg_t *args);
```
Presumably, if the priority is currently unavailable, the plugin will fire off some asynchronous work to fetch missing information, and will request reprioritization of the job in the callback for that async operation.

If a plugin job.priority callback does not return a priority, or note that priority is unavailable, then jobtap will provide a default priority to avoid jobs hanging in the PRIORITY state when a non-priority jobtap plugin is loaded.

The jobtap interface includes an installed header file `<flux/jobtap.h>` which provides some interfaces for the plugin into the job manager.  These include access to the job manager's flux_t handle for setting up timer or periodic watchers, or sending RPCs, an interface for registering plugin-specific services, and interfaces for reprioritizing and single job or requesting re-prioritization of all jobs in a loop (the plugin's `priority.set` callback will be called for each prioritized job).

Instead of defining a `flux_jobtap_t` opaque type, `jobtap.h` functions simply take a `flux_plugin_t` directly. This is as a convenience to plugin developers, as it saves a call to something like `flux_plugin_get_jobtap()` (this is done instead inside each method)

Plugins other than the default may be loaded via the job-manager.jobtap RPC. A few builtin plugins are provided for testing, including
    
 - `builtin.default` - sets priority to the current urgency (this plugin is not strictly necessary anymore and may be dropped for efficiency)
 - `builtin.age` - An experimental plugin where urgency and time since submission both contribute to priority (this one also would probably be removed before merging the PR).
 - `builtin.hold` - all jobs get an initial priority of 0 and are thus held until a non-zero urgency is set.
 - `builtin.random` - jobs are assigned a random priority. The plugin additionally calls a reprioritization loop once per second. (use for torture test)
    
The python script below can be used currently to load a non-default plugin, in case anyone wants to play with these other plugins:
    
```python
#!/usr/bin/env python3
import flux
import sys
print(flux.Flux().rpc("job-manager.jobtap", {"load": sys.argv[1]}).get())
```
    
After a new plugin is loaded, a job reprioritization loop is run, which will cause the plugin to get a priority.set callback for each job in SCHED or PRIORITY state.

#### Example

This plugin registers for all plugin topics and emits debugging info on stderr.

In the PRIORITY state it returns `flux_jobtap_priority_unavail()` and starts a timer, setting an actual priority after 1s to demonstate jobs hanging in the PRIORITY state.

```c
#include <stdio.h>
#include <jansson.h>

#include <flux/core.h>
#include <flux/jobtap.h>

struct timer_arg {
    flux_jobid_t id;
    int urgency;
    flux_plugin_t *p;
};

struct timer_arg *timer_arg_create (flux_plugin_t *p,
                                    flux_jobid_t id,
                                    int urgency)
{
    struct timer_arg *ta = calloc (1, sizeof (*ta));
    if (!ta)
        return NULL;
    ta->id = id;
    ta->p = p;
    ta->urgency = urgency;
    return ta;
}

void timer_arg_destroy (struct timer_arg *ta)
{
    free (ta);
}

static void timer_cb (flux_reactor_t *r,
                      flux_watcher_t *w,
                      int revents,
                      void *arg)
{
    struct timer_arg *ta = arg;
    flux_jobtap_reprioritize_job (ta->p, ta->id, ta->urgency);
    timer_arg_destroy (ta);
    flux_watcher_destroy (w);
}

static int cb (flux_plugin_t *p,
               const char *topic,
               flux_plugin_arg_t *args,
               void *arg)
{
    json_t *resources;
    flux_jobid_t id;
    uint32_t userid;
    int urgency;
    unsigned int priority;
    flux_job_state_t state;
    double t_submit;
    char *s;

    flux_plugin_arg_unpack (args, FLUX_PLUGIN_ARG_IN,
                            "{s:{s:o} s:I s:i s:i s:i s:i s:f}",
                            "jobspec", "resources", &resources,
                            "id", &id,
                            "userid", &userid,
                            "urgency", &urgency,
                            "priority", &priority,
                            "state", &state,
                            "t_submit", &t_submit);

    s = json_dumps (resources, JSON_COMPACT);
    fprintf (stderr, "%s: id=%ju uid=%u urg=%d pri=%u state=%d t_submit=%.5f\n",
             topic, id, userid, urgency, priority, state, t_submit);
    fprintf (stderr, "resources: %s\n", s);
    free (s);

    if (strcmp (topic, "job.priority") == 0) {
        flux_t *h = flux_jobtap_get_flux (p);
        struct timer_arg *ta = timer_arg_create (p, id, urgency);
        flux_watcher_t *w = flux_timer_watcher_create (flux_get_reactor (h),
                                                       1., 0.,
                                                       timer_cb,
                                                       ta);
        flux_watcher_start (w);
        return flux_jobtap_priority_unavail (p, args);
    }
    return 0;
}

int flux_plugin_init (flux_plugin_t *p)
{
    return flux_plugin_add_handler (p, "*", cb, NULL);
}

```

This plugin currently must be compiled via:
```console
$ gcc -I. -Isrc/include -fPIC -shared -o test.so test.c
```

At the top of the Flux build tree, but once `jobtap.h` is installed, the extra `-I. -Isrc/include` will not be necessary.


#### Benchmark

There might be some question if the addition of `jobtap_call()` to every state transition will impact maximum job throughput.

I used the following throughput tests to ensure that the impact of this PR would not be severe. The `throughput.py` test exercises the maximum rate of jobs through the job manager by submitting jobs the `exec.test` key set, which only simulates job execution (no job shell is actually launched)

These two runs compare the default plugin which will behave the same as loading no plugin:
```console
grondo@asp:~/git/flux-core.git$ src/cmd/flux start -s4 sh -c 'flux module reload sched-simple unlimited; ./throughput.py -n 1024' 2>&1 | tee output
number of jobs: 1024
script runtime: 7.846 s
job runtime:    7.661 s
throughput:     133.7 job/s (script: 130.5 job/s)

grondo@asp:~/git/flux-core.git$ src/cmd/flux start -s4 sh -c 'flux module reload sched-simple unlimited; ./jobtap.py none; ./throughput.py -n 1024' 2>&1 | tee output
{'previous': 'default', 'current': 'none'}
number of jobs: 1024
script runtime: 7.948 s
job runtime:    7.731 s
throughput:     132.5 job/s (script: 128.8 job/s)
```

It can be seen that the impact of calling a basic jobtap plugin is negligible.

Finally, the `builtin.random` plugin is a torture test for this interface, since it will reassign priorities randomly every second:

```console
$ src/cmd/flux start -s4 sh -c 'flux module reload sched-simple unlimited; ./jobtap.py builtin.random; ./throughput.py -n 1024' 2>&1 | tee output
{'previous': 'default', 'current': 'random'}
number of jobs: 1024
script runtime: 14.293s
job runtime:    14.107s
throughput:     72.6 job/s (script:  71.6 job/s)
```

This plugin seems to have a significant impact, so care will have to be taken when periodically requesting the reprioritization of all jobs.