@chu11 2021-01-11
Haven't looked through code yet, but things look great so far.  A few comments after reading the above.

> job.new: the job manager is seeing a job for the first time. This correspoonds to jobs either submitted or being reloaded from from the KVS. In the job.new callback, the job state must be explicitly checked (though it is definitely not state NEW, since the job manager does not handle jobs in this state).
> job.<state>: on transition to job state state. Note, this will not include all job states. Due to architecture of the job manager the first state plugins will see after job.new is job.priority.

I had initially assumed `job.new` meant the job state "NEW".  Perhaps a different word would be less confusing?  `job.submit`?  `job.start`???

> priority.set: when the job manager wants to recalculate priority of a job.

Perhaps its just me, but shouldn't this be `priority.get`?  The `job-manager` is asking the plugin to give it a new priority, then the `job-manager` will assign the job the new priority? 

> An interface to request reprioritization of all jobs is introduced. This causes the priority.set callback to be called for all jobs in the PRIORITY and SCHED state.

I guess if the priority plugin is doing asynchronous priority calculations, it could get a second priority request for a job before completing the first request.  There's a part of me that thinks that only jobs in SCHED state should call the callback.  But I can see the other side of it too, b/c you may want everything re-ordered after a "sweep" of all jobs.  I suppose its something that priority plugins will just have to manage and be aware can happen.  I suspect most will just do priority calculations in the order they receive requests, so it shouldn't be a problem in general.

> builtin.default - sets priority to the current urgency (this plugin is not strictly necessary anymore and may be dropped for efficiency)

Wouldn't we want to keep this?  Minimally for tests?  I guess I'm missing something about not needing this default behavior?