@grondo 2021-01-11
Good observations!

>  I had initially assumed job.new meant the job state "NEW". Perhaps a different word would be less confusing? job.submit? job.start???

Yeah, the NEW state is unfortunately named. The `job.new` plugin state event is meant to convey that this is the first the job manager is seeing of a new job, i.e. a job has been submitted (in which case the job will be in state DEPEND), or a job is being reloaded from the KVS (in which case the job can be in any state but INACTIVE I suppose). The `job.new` callback is there to allow plugins to initialize any job state if necessary (e.g. counting total number of jobs for a user, etc).

Initially there were two different callbacks: `job.depend` when a job is submitted, since the DEPEND state is the first real job state seen in the job manager, and then a separate `job.restart` callback when a job is reloaded from the KVS. However, this might make it easy for a plugin to miss one of the callbacks, thus creating difficult to diagnose bugs, so they were collapsed.

Maybe if `job.new` is no good `job.init` would work? It doesn't convey the intent of the callback as well IMO, but as long as the meaning is documented it should be fine.

> Perhaps its just me, but shouldn't this be priority.get? The job-manager is asking the plugin to give it a new priority, then the job-manager will assign the job the new priority?

You are probably right. I was thinking of these topic-based callbacks as "events" for which a plugin subscribes, so I chose `priority.set` (the job manager is asking the plugin to set a priority). However, `priority.get` works just as well for me.

> I guess if the priority plugin is doing asynchronous priority calculations, it could get a second priority request for a job before completing the first request. There's a part of me that thinks that only jobs in SCHED state should call the callback.

Good point! I had it this way at first, but  reprioritization of a jobs in PRIORITY state is necessary in the case of loading a new plugin, o/w jobs will be stuck in that state. As an optimization we could perhaps only re-prioritize jobs in SCHED during a normal re-prioritization loop, but it seemed easy enough for a plugin which allows jobs to stick in PRIORITY state to just not return a priority in this case (thereby keeping jobs waiting on priority in PRIORITY state). Also for plugins which calculate priorities on an interval, it seemed like a way to enable a plugin to only release jobs from PRIORITY state on that interval.

> Wouldn't we want to keep this? Minimally for tests? I guess I'm missing something about not needing this default behavior?

Sorry, I may not have been clear. The default behavior is now the same as the builtin default when _no_ plugin is loaded, so it is as least not necessary to load `builtin.default` by default.