@chu11 2021-01-11
> Maybe if job.new is no good job.init would work? It doesn't convey the intent of the callback as well IMO, but as long as the meaning is documented it should be fine.

I'll throw out `job.create` (created?) as well as an option.  Or alternately something like `job.state-<state>`, so it's more clear that `job.new` is not the same as `job.state-new`.