@grondo 2021-01-11
> Or alternately something like job.state-<state>, so it's more clear that job.new is not the same as job.state-new.

In a previous version, the state transition callbacks were `state.<state>` while the restart callback was `job.restart`. Maybe I can go back to that scheme, but instead of `job.restart` keep the `job.new` hook for jobs during restart, as well as the one when the job manager first sees a job and transitions it from `SUBMIT` to `DEPEND`.

Looking forward, the job manager could offer other, non-job callback topics. In that case, maybe it is indeed prudent to keep all the current callbacks named with a `job.` prefix. In that case, maybe `job.state-<state>` as you suggested, or `job.state.<state>` to keep all the "state transition" topics organized.

If we take this approach, then maybe `priority.get` should be renamed `job.get-priority`.